src/agents/system_architect.py
"""System Architect Agent - Simplified Version"""

import autogen
from typing import Dict, Any, List
import logging
from datetime import datetime


class SystemArchitectAgent(autogen.AssistantAgent):
    """System Architect Agent - Simplified for testing"""
    
    def __init__(self, name: str = "system_architect", **kwargs):
        system_message = "You are a System Architect Agent at ConsultingAI Digital Advisory Firm."
        
        super().__init__(
            name=name,
            system_message=system_message,
            llm_config=False,
            human_input_mode="NEVER",
            **kwargs
        )
        
        self.capabilities = [
            "system_design",
            "scalability_analysis", 
            "integration_planning",
            "technology_selection",
            "api_design",
            "infrastructure_planning"
        ]
        
        self.architecture_history = []
        self.logger = logging.getLogger(f"ConsultingAI.agents.{name}")
        
        print(f"✅ {name} created successfully")
    
    def analyze_system_architecture(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """Simple architecture analysis"""
        
        analysis_id = f"arch_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
        
        # Mock analysis results
        result = {
            "analysis_id": analysis_id,
            "timestamp": datetime.now().isoformat(),
            "agent": self.name,
            "context": context,
            "confidence": 0.8,
            "recommendations": [
                {"category": "scalability", "recommendation": "Use microservices", "priority": "high"},
                {"category": "performance", "recommendation": "Implement caching", "priority": "medium"}
            ],
            "technology_stack": {
                "backend": "FastAPI",
                "database": "PostgreSQL", 
                "caching": "Redis",
                "deployment": "Docker",
                "monitoring": "Prometheus"
            }
        }
        
        self.architecture_history.append(result)
        return result
    
    def get_agent_capabilities(self) -> Dict[str, Any]:
        """Get agent capabilities"""
        return {
            "agent_type": "SystemArchitectAgent",
            "name": self.name,
            "capabilities": self.capabilities,
            "total_analyses": len(self.architecture_history)
        }


def create_system_architect_agent(**kwargs):
    """Factory function"""
    return SystemArchitectAgent(**kwargs)


def demonstrate_system_architect():
    """Demonstration function"""
    print("🔧 Demonstrating System Architect Agent functionality...")
    
    try:
        # Create agent
        architect = create_system_architect_agent(name="demo_system_architect")
        
        # Test analysis
        context = {
            "type": "web_application",
            "scale": "high",
            "complexity": "medium"
        }
        
        result = architect.analyze_system_architecture(context)
        
        print(f"  ✅ Architecture analysis completed: {result['analysis_id']}")
        print(f"     Confidence: {result['confidence']:.1%}")
        print(f"     Recommendations: {len(result['recommendations'])} items")
        print(f"     Technology stack: {len(result['technology_stack'])} components")
        
        # Test capabilities
        capabilities = architect.get_agent_capabilities()
        print(f"  ✅ Agent capabilities: {len(capabilities['capabilities'])} core capabilities")
        
        return True
        
    except Exception as e:
        print(f"  ❌ System Architect demonstration failed: {e}")
        return False


if __name__ == "__main__":
    print("🚀 Starting SystemArchitectAgent demonstration...")
    
    success = demonstrate_system_architect()
    if success:
        print("\n✅ System Architect Agent: DEMONSTRATED")
    else:
        print("\n❌ System Architect Agent: FAILED")
        exit(1)
