pyproject.toml
[project]
name = "consultingai"
version = "0.1.0"
description = "A sophisticated multi-agent system implementing Microsoft AutoGen's Society of Mind framework through consulting industry patterns"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogen",
    "pydantic",
    "pyyaml",
    "pytest",
    "python-dotenv>=1.1.1",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.0",
    "pylint>=3.3.7",
    "pytest-cov>=6.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[tool.pylint.messages_control]
disable = ["C0114", "C0116"]  # missing docstrings

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20"
]
